Add files via upload
When the right key is pressed have the 'audio_flyoff' sound play as a one shot and have the sprite_spaceship fly off the right side of the screen with out any stopping. Do not destroy the sprite as it may be used later.

If sprite_spaceship is not on screen then when the spacebar is pressed and released  play the audio file "rocket-launch fly-off mono 8khz 1s.mp3" as a one shot and make the sprite fly in from the left side of the screen and stop at the center of the screen.